move tests for sanitize_tri to test_utils
